Split out the 15min tool to its own crate.
