+++ book/src/dev/README.md
@@ -142,6 +142,7 @@ Other:
 - `experiment`: ...
 - `parking_mapper`: a standalone tool to help map street parking in OSM
 - `osm_viewer`: a standalone tool to render OSM in detail
+- `fifteen_min`: a standalone tool to explore 15-minute neighborhoods
 
 ## Code conventions
