+++ game/src/layer/mod.rs
@@ -1,9 +1,9 @@
+use map_gui::common::HeatmapOptions;
 use widgetry::{
     Btn, DrawBaselayer, EventCtx, GfxCtx, Key, Line, Outcome, Panel, State, TextExt, Widget,
 };
 
 use crate::app::App;
-use crate::common::HeatmapOptions;
 use crate::game::Transition;
 use crate::helpers::{grey_out_map, hotkey_btn};
 use crate::sandbox::dashboards;
